Move Vue.php
